FEAT: Added global exception handler
# Conflicts:
#	src/main/java/com/bicart/controller/SubCategoryController.java
#	src/main/java/com/bicart/service/CartService.java
#	src/main/java/com/bicart/service/CategoryService.java
#	src/main/java/com/bicart/service/OrderItemService.java
#	src/main/java/com/bicart/service/OrderService.java
#	src/main/java/com/bicart/service/PaymentService.java
#	src/main/java/com/bicart/service/ProductService.java
#	src/main/java/com/bicart/service/RoleService.java
#	src/main/java/com/bicart/service/UserService.java

+++ src/main/java/com/bicart/dto/ProductDto.java
@@ -18,7 +18,7 @@ public class ProductDto {
     private String description;
 
     @NotBlank(message = "Product must have price")
-    private double price;
+    private long price;
 
     @NotBlank(message = "Product must have remaining stock")
     private int quantity;
